Clean up pre-existing clippy lints in host and client
- client: drop a redundant i64->i64 cast in the hand-rolled RFC-3339 formatter.
- host: annotate build_source with allow(clippy::needless_return) — each cfg
  arm must return because exactly one survives per target and the others leave
  no tail expression; clippy only sees the active arm.

Workspace is now clippy-clean (cargo clippy --workspace --all-targets) with
all tests passing.

+++ simulboot-client/src/session/mod.rs
@@ -82,7 +82,7 @@ fn format_epoch_utc(secs: u64) -> String {
     // days since 1970-01-01 -> civil date
     let z = days + 719_468;
     let era = if z >= 0 { z } else { z - 146_096 } / 146_097;
-    let doe = (z - era * 146_097) as i64; // [0, 146096]
+    let doe = z - era * 146_097; // [0, 146096]
     let yoe = (doe - doe / 1460 + doe / 36524 - doe / 146096) / 365; // [0, 399]
     let y = yoe + era * 400;
     let doy = doe - (365 * yoe + yoe / 4 - yoe / 100); // [0, 365]
